update quantization filters to handle tensor
